encoding/json: expose efficient IsValid()
There's a lovely, allocation-free function for scanning a byte array to
see if it's valid JSON tucked away inside the encoding/json package, and
used in several places within the package.

This just exposes it as a public function with a couple of tests.

Hopefully the scanner allocation here is optimised onto the stack and
this will also avoid heap churn.

Change-Id: Ie043ad4c1e3c934a43adb0c6c09c287c0514c6fa